requirements.txt
#
# Production requirements
#

alembic
babel
celery[redis]>=3.1,<3.1.99
cssmin
filemagic
gevent
gunicorn
humanize
lingua
psycopg2
python-dateutil
python-slugify
pyramid>=1.7
pyramid_chameleon
pyramid_exclog
pyramid_tm==0.12.1
pyramid_redis_sessions
pyramid_redis
pyramid_webassets
pyramid_who
repoze.who>=2.3.0
six
SQLAlchemy>=1.0.0,<1.1.0
tabulate
wtforms>=2.0.0
wtforms-json
wtforms-components
zope.sqlalchemy
